split use_args decorator in seprate file to support python 3.4
